Java Supermarket Basket System
A relatively simple program that helps the user to keep in mind the products he or she wants to buy from
the local supermarket. It utilizes a basket that has 50 slots for the products to be added. If the user decides to
delete or see the available products then this utility is available.