Java Text Management System
A text management system written in java, the program is useful especially when someone wants to get statistically in a percentage formation every character of the text file he imposed the system to read. The system reads every character the text given contains and initializes a doubly linked list which uses a hashmap to store the data. In addition the system was designed from the ground up to hold the times a character found in a class member variable and also in the data hashmap structure.