Java Supermarket Basket System
A java programmed system that utilizes a basket system for anyone to be helped to avoid forgetting his supermarket basket list. The program has 50 available entries to be stored and the user can edit, delete, insert and view all of his entries. A simple but useful program that anyone can use!